Live session injection via claude-commander + markdown response toast
Add claude-commander integration so the iPad can inject messages (text + images)
into a running interactive Claude Code session via Unix socket. Includes launcher
script, backend injection module, new API endpoints, subagents for parallel
draw/widget execution, hooks for idle detection, and markdown rendering in the
iPad response toast.

+++ iPad/iris-app/Views/ToolbarView.swift
@@ -35,8 +35,8 @@ private struct OpenAILogo: Shape {
         // Outer hexagon
         for i in 0..<6 {
             let angle = Angle.degrees(Double(i) * 60 - 90)
-            let x = cx + cos(angle.radians) * s * 0.48
-            let y = cy + sin(angle.radians) * s * 0.48
+            let x = cx + CGFloat(cos(angle.radians)) * s * 0.48
+            let y = cy + CGFloat(sin(angle.radians)) * s * 0.48
             if i == 0 { p.move(to: CGPoint(x: x, y: y)) }
             else { p.addLine(to: CGPoint(x: x, y: y)) }
         }
@@ -46,8 +46,8 @@ private struct OpenAILogo: Shape {
             let angle = Angle.degrees(Double(i) * 60 - 90)
             p.move(to: CGPoint(x: cx, y: cy))
             p.addLine(to: CGPoint(
-                x: cx + cos(angle.radians) * s * 0.32,
-                y: cy + sin(angle.radians) * s * 0.32
+                x: cx + CGFloat(cos(angle.radians)) * s * 0.32,
+                y: cy + CGFloat(sin(angle.radians)) * s * 0.32
             ))
         }
         return p
